Fix larger attachments not received
12 second timeout (over default 3) arrived at by brief experimentation:
an email with ~24MB of attachments (2) was rejected by SES (and so never
reached the smtp-relay Lambda), while another with ~11.5MB attachment
was processed locally with the test container in a hair over 6 seconds.

Real Lambda may get less CPU, (its proportional to RAM, currently at the
min. 128MB) but it also gets faster network I/O to S3. Most emails are
relayed in ~1.7s.

A good future improvement would be to set a shorter timeout (would have
to be in the handler code) if the spam verdict is 'fail'.