Fix edited address headers having no effect
Previously edits to To, Bcc etc. addresses would be lost on the
confirmation dialogue (and from the subsequently sent message) since the
replyMeta object was constructed from the object returned from the reply
template fn (which differed from the EmlMeta type used elsewhere and
needed for the addresses) but changes were not propagated back.

The reply template now returns the same EmlMeta object, so this can be
used and edited directly.